+++ transmart-data/ddl/postgres/tm_wz/wt_subject_rbm_calcs.sql
@@ -4,9 +4,9 @@
 CREATE TABLE wt_subject_rbm_calcs (
     trial_name character varying(50),
     probeset_id character varying(1000),
-    mean_intensity bigint,
-    median_intensity bigint,
-    stddev_intensity bigint
+    mean_intensity double precision,
+    median_intensity double precision,
+    stddev_intensity double precision
 );
 
 --
